move map macros
